requirements.txt
streamlit==1.22.0
skia-python==87.5
opencv-python-headless==4.7.0.72
numpy==1.23.5
tensorflow==2.12
Pillow==9.5.0
